+++ info.h
@@ -26,6 +26,7 @@
 
 // Needed for action function pointer handling.
 #include "d_think.h"
+#include <stdint.h>
 
 typedef enum
 {
@@ -1147,12 +1148,12 @@ typedef enum
 typedef struct
 {
   spritenum_t	sprite;
-  long			frame;
-  long			tics;
+  int32_t			frame;
+  int32_t			tics;
   // void		(*action) ();
   actionf_t			action;
   statenum_t			nextstate;
-  long			misc1, misc2;
+  int32_t			misc1, misc2;
 } state_t;
 
 extern state_t	states[NUMSTATES];
